refactor(ComponentDoc): refactor to HOC
