+++ docs/app/Components/ComponentDoc/index.js
@@ -1,0 +1,1 @@
+export default from './ComponentDoc'
