+++ docs/app/HOC/index.js
@@ -1,3 +1,4 @@
 export neverUpdate from './neverUpdate'
 export pure from './pure'
 export updateForKeys from './updateForKeys'
+export withDocInfo from './withDocInfo'
