+++ docs/app/utils/index.js
@@ -4,6 +4,8 @@ import * as semanticUIReact from 'src'
 import { META } from 'src/lib'
 
 export * from './constants'
+export getComponentGroup from './getComponentGroup'
+export getSeeItems from './getSeeItems'
 export scrollToAnchor from './scrollToAnchor'
 
 /**
